Removed the flowchart because it was a failure.
Taking a break from this task for today. I had expected problems with createBitmapImage because it had a weird signature. And that's where everything crashed.

On the surface, it seems like I need to learn about Promise. But I think that what I'm missing is the skills to deal with asynchronous execution.

GG

See you again later, JavaScript